big refactor in structure
