+++ components/ImageUpload/UploadStyles.js
@@ -1,0 +1,12 @@
+import { StyleSheet } from 'react-native';
+
+export const UploadStyles = StyleSheet.create({
+  imageUpload: {
+    height: 120,
+    width: 120,
+    borderRadius: 60,
+    borderColor: 'grey',
+    borderWidth: 1,
+    padding: 30,
+  }
+})
