+++ components/headers/Header.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import { Text } from 'react-native'
-import { HeaderStyles } from '../../styles/HeaderStyles';
+import { HeaderStyles } from './HeaderStyles';
 
 
 const Header = ({text}) => {
