+++ pages/league/LeaguePage.js
@@ -1,9 +1,9 @@
 import React, { Component } from 'react'
 import { View, Text, Animated, Easing, Image } from 'react-native'
 import NewLeague from './NewLeague'
-import { HomeStyle } from '../../styles/HomeStyle';
+import { ViewStyles } from '../../styles/ViewStyles';
 import { AppStyle } from '../../styles/AppStyle';
-import { CardStyle } from '../../styles/CardStyle';
+import { CardStyle } from '../../components/cards/CardStyle';
 import { connect } from 'react-redux'
 import { setLeague } from '../../actions/leagueActions'
 import Header from '../../components/headers/Header'
@@ -15,9 +15,9 @@ class LeaguePage extends Component {
       const { name, image_url, description, number_of_teams } = this.props.currentLeague
 
       return (
-        <View style={ HomeStyle.profile } >
-          <View style={ AppStyle.firstLayer } >
-            <View style={HomeStyle.userCard} >
+        <View style={ ViewStyles.profile } >
+          <View style={ ViewStyles.firstLayer } >
+            <View style={CardStyle.userCard} >
               <Image
                 source={{uri: image_url }}
                 style={[AppStyle.avatar, { height: 150, width: 150}, { borderRadius: 75 }]} />
@@ -25,7 +25,7 @@ class LeaguePage extends Component {
               <Header text={description} />
             </View>
           </View>
-          <View style={ HomeStyle.secondLayer } >
+          <View style={ ViewStyles.secondLayer } >
             <Header text="This week's workouts"/>
             <Text>so many workouts</Text>
           </View>
@@ -38,7 +38,7 @@ class LeaguePage extends Component {
 
   render(){
     return(
-      <View style={ HomeStyle.profile } >
+      <View style={ ViewStyles.profile } >
         {this.renderLeague()}
       </View>
     )
