+++ styles/ViewStyles.js
@@ -1,0 +1,32 @@
+import { StyleSheet } from 'react-native';
+
+export const ViewStyles = StyleSheet.create({
+  profile: {
+    flex: 1,
+    backgroundColor: '#D5DDE7'
+  },
+  firstLayer: {
+    flex: 1,
+    paddingTop: 80,
+    backgroundColor: '#D5DDE7'
+  },
+  secondLayer: {
+    backgroundColor: '#B5C9D3',
+    borderTopLeftRadius: 15,
+    borderTopEndRadius: 15,
+    paddingBottom: 20,
+    marginBottom: -15,
+  },
+  thirdLayer: {
+    backgroundColor: '#B994D8',
+    borderTopLeftRadius: 15,
+    borderTopEndRadius: 15,
+  },
+  signUpPage: {
+    flex: 1,
+    padding: 10,
+    display: 'flex',
+    flexDirection: 'column',
+    alignItems: 'center'
+  },
+})
